[PATCH] asm-arm26/hdreg.h: use unsigned long for ide_ioreg_t

+++ include/asm-arm26/hdreg.h
@@ -9,7 +9,7 @@
 #ifndef __ASMARM_HDREG_H
 #define __ASMARM_HDREG_H
 
-typedef unsigned int ide_ioreg_t;
+typedef unsigned long ide_ioreg_t;
 
 #endif /* __ASMARM_HDREG_H */
